# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nodeticket API",
    "description": "RESTful API for Nodeticket help desk system. Backward compatible with osTicket v1.8+.",
    "version": "1.0.0",
    "contact": {
      "name": "Nodeticket Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/staff/{id}": {
      "get": {
        "tags": ["Staff"],
        "summary": "Get staff member details",
        "operationId": "getStaff",
        "security": [{"bearerAuth": []}, {"apiKey": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
    "schemas": {
...
      "Staff": {
        "type": "object",
        "properties": {
          "staff_id": {"type": "integer"},
          "dept_id": {"type": "integer"},
          "role_id": {"type": "integer"},
          "username": {"type": "string"},
          "firstname": {"type": "string"},
          "lastname": {"type": "string"},
          "email": {"type": "string"},
          "phone": {"type": "string"},
          "isactive": {"type": "boolean"},
          "isadmin": {"type": "boolean"},
          "onvacation": {"type": "boolean"},
          "created": {"type": "string", "format": "date-time"}
        }
      },
...
      "StaffDetail": {
        "allOf": [
          {"$ref": "#/components/schemas/Staff"},
          {
            "type": "object",
            "properties": {
              "department": {"$ref": "#/components/schemas/DepartmentSummary"},
              "role": {"$ref": "#/components/schemas/Role"},
              "teams": {
                "type": "array",
                "items": {"$ref": "#/components/schemas/TeamSummary"}
              },
              "departments": {
                "type": "array",
                "items": {"$ref": "#/components/schemas/DepartmentSummary"}
              }
            }
          }
        ]
      },
...
      "DepartmentSummary": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"}
        }
      },
...
      "TeamSummary": {
        "type": "object",
        "properties": {
          "team_id": {"type": "integer"},
          "name": {"type": "string"}
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "permissions": {"type": "object"}
        }
      },
...
    }
  }
}